refactored views and added new Environment files and view extension files
