+++ Sources/Dragula/Extensions/View+Hidden.swift
@@ -2,8 +2,7 @@
 //  View+Hidden.swift
 //  Dragula
 //
-//  Created by Eugene Kovs on 04.07.2025.
-//  https://github.com/kovs705
+//  Refactored by Eugene Kovs on 04.07.2025.
 //  Contributed to Dragula
 //
